// apps/sales-copilot/src/generated/services/Biz_appfeedbacksService.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */

import type { Biz_appfeedbacksBase, Biz_appfeedbacks } from '../models/Biz_appfeedbacksModel';
import type { GetEntityMetadataOptions, EntityMetadata } from '@microsoft/power-apps/data/metadata/dataverse';
import type { IGetOptions, IGetAllOptions } from '../models/CommonModels';
import type { IOperationResult } from '@microsoft/power-apps/data';
import { dataSourcesInfo } from '../../../.power/schemas/appschemas/dataSourcesInfo';
import { getClient } from '@microsoft/power-apps/data';


export class Biz_appfeedbacksService {
  private static readonly dataSourceName = 'biz_appfeedbacks';

  private static readonly client = getClient(dataSourcesInfo);

  public static async create(record: Omit<Biz_appfeedbacksBase, 'biz_appfeedbackid'>): Promise<IOperationResult<Biz_appfeedbacks>> {
    const result = await Biz_appfeedbacksService.client.createRecordAsync<Omit<Biz_appfeedbacksBase, 'biz_appfeedbackid'>, Biz_appfeedbacks>(
      Biz_appfeedbacksService.dataSourceName,
      record
    );
    return result;
  }

  public static async update(id: string, changedFields: Partial<Omit<Biz_appfeedbacksBase, 'biz_appfeedbackid'>>): Promise<IOperationResult<Biz_appfeedbacks>> {
    const result = await Biz_appfeedbacksService.client.updateRecordAsync<Partial<Omit<Biz_appfeedbacksBase, 'biz_appfeedbackid'>>, Biz_appfeedbacks>(
      Biz_appfeedbacksService.dataSourceName,
      id,
      changedFields
    );
    return result;
  }

  public static async delete(id: string): Promise<void> {
    await Biz_appfeedbacksService.client.deleteRecordAsync(
      Biz_appfeedbacksService.dataSourceName,
      id);
  }

  public static async get(id: string, options?: IGetOptions): Promise<IOperationResult<Biz_appfeedbacks>> {
    const result = await Biz_appfeedbacksService.client.retrieveRecordAsync<Biz_appfeedbacks>(
      Biz_appfeedbacksService.dataSourceName,
      id,
      options
    );
    return result;
  }

  public static async getAll(options?: IGetAllOptions): Promise<IOperationResult<Biz_appfeedbacks[]>> {
    const result = await Biz_appfeedbacksService.client.retrieveMultipleRecordsAsync<Biz_appfeedbacks>(
      Biz_appfeedbacksService.dataSourceName,
      options
    );
    return result;
  }

  public static getMetadata(
    options: GetEntityMetadataOptions<Biz_appfeedbacks> = {}
  ): Promise<IOperationResult<Partial<EntityMetadata>>> {
    return Biz_appfeedbacksService.client.executeAsync({
      dataverseRequest: {
        action: "getEntityMetadata",
        parameters: {
          tableName: Biz_appfeedbacksService.dataSourceName,
          options: options as GetEntityMetadataOptions,
        },
      },
    });
  }
}