// apps/sales-copilot/src/generated/models/Biz_appfeedbacksModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Biz_appfeedbacksstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Biz_appfeedbacksstatecode = keyof typeof Biz_appfeedbacksstatecode;
export const Biz_appfeedbacksstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Biz_appfeedbacksstatuscode = keyof typeof Biz_appfeedbacksstatuscode;

export interface Biz_appfeedbacksBase {
  biz_appfeedbackid: string;
  biz_appversion: string;
  biz_browser?: string;
  biz_buildid: string;
  biz_clientrequestid: string;
  biz_currentpage?: string;
  biz_description: string;
  biz_device?: string;
  biz_expectedoutcome?: string;
  biz_feedbacktype: string;
  biz_githubissuenumber?: number;
  biz_githubissueurl?: string;
  biz_locale: string;
  biz_os?: string;
  biz_reproductionsteps?: string;
  biz_source: string;
  biz_submissionstatus: string;
  biz_submittedon: string;
  biz_syncerror?: string;
  biz_title: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Biz_appfeedbacksstatecode;
  statuscode?: Biz_appfeedbacksstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Biz_appfeedbacks extends Biz_appfeedbacksBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}